requirements.txt
blis==0.7.4
catalogue==2.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.0
MarkupSafe==1.1.1
murmurhash==1.0.5
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pathy==0.3.5
preshed==3.0.5
psycopg2-binary==2.8.6
pydantic==1.7.3
pyparsing==2.4.7
PySocks==1.7.1
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
smart-open==3.0.0
spacy==3.0.1
spacy-legacy==3.0.1
SQLAlchemy==1.3.23
srsly==2.4.0
thinc==8.0.1
threadpoolctl==2.1.0
tqdm==4.56.0
tweepy==3.10.0
typer==0.3.2
urllib3==1.26.3
wasabi==0.8.2
Werkzeug==1.0.1
